Refactor: Modularize codebase into logical components
- Created new directory structure (utils, api, handlers, objects)
- Moved error handling to a separate utility module
- Extracted API client functionality to its own module
- Created separate modules for resource and tool handlers
- Moved company-specific functionality to a dedicated module
- Updated index.ts to use the new modules
- Fixed import paths for ES modules
