+++ package.json
@@ -21,17 +21,17 @@
   ],
   "dependencies": {
     "@modelcontextprotocol/sdk": "^1.4.1",
-    "axios": "^1.7.9",
-    "shx": "^0.3.4",
-    "typescript": "^5.7.2"
+    "axios": "^1.7.9"
   },
   "author": "@hmk",
   "license": "BSD-3-Clause",
   "devDependencies": {
     "@types/jest": "^29.5.14",
     "jest": "^29.7.0",
+    "shx": "^0.4.0",
     "ts-jest": "^29.2.5",
-    "tsx": "^4.19.2"
+    "tsx": "^4.19.2",
+    "typescript": "^5.8.3"
   },
   "jest": {
     "preset": "ts-jest",
